+++ CHANGES.rst
@@ -9,6 +9,7 @@ in progress
 - More detailed logging of post details while processing posts
 - Adjust margin size to 2cm when converting with pandoc
 - Rename "--path" parameter to "--output-path"
+- Propagate topic creation/update date to pandoc metadata
 
 
 2019-08-12 0.2.0